backend_scaffold/requirements.txt
fastapi[all]
uvicorn
opencv-python-headless
numpy
tensorflow-cpu
google-generativeai
python-multipart
python-dotenv
